+++ packages/terra-clinical-item-view/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added
   * Added `trueColumn` prop to toggle between displaying in a two column layout by row or by column.
 
+* Changed
+  * Changed the structure of how displays in an item view are returned, now they exist in an unordered list as list items.
+
 ## 4.11.0 - (September 21, 2023)
 
 * Changed