+++ packages/terra-clinical-item-view/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added
   * Added `trueColumn` prop to toggle between displaying in a two column layout by row or by column.
+  * Added accessibility guide.
 
 ## 4.11.0 - (September 21, 2023)